tracker-app/test_adaptive_colors_real.py
#!/usr/bin/env python3
"""
Test adaptive color learning using REAL images from the matches database.
Tests with actual toddler images that may have same/different shirt colors.
"""

import sqlite3
import numpy as np
import cv2
import sys
from datetime import datetime, timedelta
from io import BytesIO

# Import the actual modules
from hybrid_identifier import ColorMatcher

def load_images_from_db(db_path, subject="Erik", limit=10):
    """Load real images from matches database."""
    conn = sqlite3.connect(db_path)
    cursor = conn.cursor()

    cursor.execute("""
        SELECT image_data, timestamp, confidence
        FROM matches
        WHERE subject = ?
        ORDER BY timestamp DESC
        LIMIT ?
    """, (subject, limit))

    images = []
    for image_data, timestamp, confidence in cursor.fetchall():
        if image_data:
            # Decode image from database blob
            nparr = np.frombuffer(image_data, np.uint8)
            img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
            if img is not None:
                images.append({
                    'image': img,
                    'timestamp': timestamp,
                    'confidence': confidence
                })

    conn.close()
    return images

def test_real_color_learning():
    """Test color learning with real toddler images."""
    print("="*70)
    print("ADAPTIVE COLOR LEARNING TEST - REAL IMAGES")
    print("="*70)

    # Load real images
    db_path = "matches.db"
    print(f"\nLoading images from {db_path}...")

    try:
        images = load_images_from_db(db_path, subject="Erik", limit=10)
    except Exception as e:
        print(f"Error loading images: {e}")
        return False

    if len(images) < 2:
        print(f"Not enough images found. Need at least 2, got {len(images)}")
        return False

    print(f"✓ Loaded {len(images)} real images of Erik")

    # Initialize color matcher
    matcher = ColorMatcher()

    # Test 1: Extract colors from first image (simulating first detection of day)
    print("\n" + "-"*70)
    print("TEST 1: Learning colors from first detection")
    print("-"*70)

    first_image = images[0]['image']
    print(f"First image size: {first_image.shape}")

    first_colors = matcher.extract_color_features(first_image)
    print(f"✓ Extracted color features: {first_colors.shape}")

    # Store as today's colors (simulating high-confidence OSNet+Face match)
    matcher.update_daily_colors(
        person_id="Erik",
        color_features=first_colors,
        confidence=0.85,
        source_methods=['osnet', 'face']
    )
    print("✓ Stored daily colors for Erik")

    # Test 2: Try to match subsequent images
    print("\n" + "-"*70)
    print("TEST 2: Matching subsequent detections")
    print("-"*70)

    match_count = 0
    no_match_count = 0
    confidences = []

    for i, img_data in enumerate(images[1:6], 1):  # Test next 5 images
        img = img_data['image']
        img_colors = matcher.extract_color_features(img)

        result = matcher.identify_person_by_color(img_colors, threshold=0.5)

        if result:
            person_id, confidence = result
            match_count += 1
            confidences.append(confidence)
            status = "✓ MATCH"
        else:
            no_match_count += 1
            confidence = 0.0
            status = "✗ NO MATCH"

        print(f"  Image {i}: {status} (confidence: {confidence:.2f})")

    print(f"\nResults:")
    print(f"  Matches: {match_count}/5")
    print(f"  No matches: {no_match_count}/5")
    if confidences:
        print(f"  Avg match confidence: {np.mean(confidences):.2f}")
        print(f"  Min/Max confidence: {np.min(confidences):.2f} / {np.max(confidences):.2f}")

    # Test 3: Verify different person would not match
    print("\n" + "-"*70)
    print("TEST 3: Color freshness and expiration")
    print("-"*70)

    assert matcher.has_fresh_colors("Erik"), "Should have fresh colors"
    print("✓ Colors are fresh (< 12 hours old)")

    # Check cache status
    cache_entry = matcher.daily_color_cache["Erik"]
    age_seconds = (datetime.now() - cache_entry['timestamp']).total_seconds()
    print(f"✓ Color cache age: {age_seconds:.1f} seconds")
    print(f"✓ Source methods: {cache_entry['source_methods']}")
    print(f"✓ Confidence: {cache_entry['confidence']:.2f}")

    # Simulate expiration
    old_time = datetime.now() - timedelta(hours=13)
    cache_entry['timestamp'] = old_time

    assert not matcher.has_fresh_colors("Erik"), "Should be expired"
    print("✓ Colors correctly expire after 12 hours")

    # Test 4: Clear and verify
    print("\n" + "-"*70)
    print("TEST 4: Cache management")
    print("-"*70)

    matcher.clear_expired_colors()
    assert "Erik" not in matcher.daily_color_cache, "Should be removed"
    print("✓ Expired colors cleared from cache")

    # Re-learn from a different image (simulating new day)
    matcher.update_daily_colors(
        person_id="Erik",
        color_features=matcher.extract_color_features(images[2]['image']),
        confidence=0.90,
        source_methods=['face']
    )
    print("✓ Re-learned colors from different image (new day)")

    assert matcher.has_fresh_colors("Erik"), "Should have fresh colors again"
    print("✓ New colors are fresh")

    # Clear all
    matcher.clear_all_colors()
    assert len(matcher.daily_color_cache) == 0, "Should be empty"
    print("✓ Cleared all colors successfully")

    return True

def test_color_similarity_variations():
    """Test how color similarity varies with real images."""
    print("\n" + "="*70)
    print("COLOR SIMILARITY ANALYSIS - REAL IMAGES")
    print("="*70)

    try:
        images = load_images_from_db("matches.db", subject="Erik", limit=10)
    except Exception as e:
        print(f"Error loading images: {e}")
        return False

    if len(images) < 3:
        print("Not enough images for similarity analysis")
        return False

    matcher = ColorMatcher()

    # Extract colors from all images
    print(f"\nExtracting colors from {len(images)} images...")
    color_features = []
    for img_data in images:
        colors = matcher.extract_color_features(img_data['image'])
        color_features.append(colors)

    print("✓ Extracted all color features")

    # Compute pairwise similarities
    print("\nPairwise color similarity matrix:")
    print("(1.00 = identical, 0.00 = completely different)")
    print()
    print("     ", end="")
    for i in range(min(5, len(images))):
        print(f"  Img{i+1}", end="")
    print()

    for i in range(min(5, len(images))):
        print(f"Img{i+1}", end=" ")
        for j in range(min(5, len(images))):
            similarity = matcher.compute_color_similarity(
                color_features[i],
                color_features[j]
            )
            print(f" {similarity:.2f}", end=" ")
        print()

    # Analyze similarity distribution
    similarities = []
    for i in range(len(color_features)):
        for j in range(i+1, len(color_features)):
            sim = matcher.compute_color_similarity(color_features[i], color_features[j])
            similarities.append(sim)

    if similarities:
        print(f"\nSimilarity Statistics:")
        print(f"  Mean: {np.mean(similarities):.2f}")
        print(f"  Std Dev: {np.std(similarities):.2f}")
        print(f"  Min: {np.min(similarities):.2f}")
        print(f"  Max: {np.max(similarities):.2f}")
        print(f"  Median: {np.median(similarities):.2f}")

        # Threshold analysis
        threshold = 0.5
        above_threshold = sum(1 for s in similarities if s >= threshold)
        print(f"\n  Images above {threshold} threshold: {above_threshold}/{len(similarities)}")
        print(f"  Percentage: {100*above_threshold/len(similarities):.1f}%")

    return True

def main():
    """Run all tests."""
    print("\n" + "="*70)
    print("ADAPTIVE COLOR LEARNING - REAL IMAGE TEST SUITE")
    print("="*70)

    try:
        success1 = test_real_color_learning()
        success2 = test_color_similarity_variations()

        if success1 and success2:
            print("\n" + "="*70)
            print("ALL TESTS PASSED ✓✓✓")
            print("="*70)
            print("\nKey Findings:")
            print("  ✓ Color learning works with real toddler images")
            print("  ✓ Same-day detections show varying similarity")
            print("  ✓ Color cache expires correctly after 12 hours")
            print("  ✓ Cache management functions work as expected")
            print("\nREADY FOR PRODUCTION!")
        else:
            print("\n⚠ Some tests could not complete (may need more images)")

    except Exception as e:
        print(f"\n❌ ERROR: {e}")
        import traceback
        traceback.print_exc()
        sys.exit(1)

if __name__ == "__main__":
    main()
